moving functions to IsoWorld
