+++ src/positionable/IsometricMap.cpp
@@ -50,7 +50,9 @@ void IsometricMap::_enter_tree() {
 
 void IsometricMap::_process(float delta) {
     //if(Engine::get_singleton()->is_editor_hint()) to check if in editor, to deactivate code if needed
-    generateTopologicalRenderGraph();
+    if( world && worldOwner) {
+        world->generateTopologicalRenderGraph();
+    }
 }
 
 String IsometricMap::get_class() const {
@@ -71,61 +73,6 @@ void IsometricMap::onGridUpdated(int stair) {
             isometricPositionable->onGridUpdated(stair - static_cast<int>(getLocal3DPosition().z));
         }
     }
-}
-
-void IsometricMap::generateTopologicalRenderGraph() {
-    if( world && worldOwner) {
-        currentSortingOrder = 0;
-        const Array &positionables = world->getPositionables();
-        for (int i = 0; i < positionables.size(); i++) {
-            auto *positionable = cast_to<IsometricPositionable>(positionables[i]);
-            if (positionable) {
-                positionable->setRendered(false);
-            }
-        }
-        for (int i = 0; i < positionables.size(); i++) {
-            auto *positionable = cast_to<IsometricPositionable>(positionables[i]);
-            if (positionable) {
-                if (!positionable->isRendered()) {
-                    renderIsoNode(positionable);
-                }
-            }
-        }
-    }
-}
-
-void IsometricMap::renderIsoNode(IsometricPositionable *isoNode) {
-    isoNode->setRendered(true);
-    int maxZSize { 0 };
-    const Array &isoNodes { getPositionableBehind(isoNode) };
-    for (int i = 0; i < isoNodes.size(); i++) {
-        auto *positionable = cast_to<IsometricPositionable>(isoNodes[i]);
-        if (positionable) {
-            if (!positionable->isRendered()) {
-                renderIsoNode(positionable);
-            }
-            int positionableZOrderSize { positionable->getZOrderSize() };
-            maxZSize = positionableZOrderSize >= maxZSize ? positionableZOrderSize : maxZSize;
-        }
-    }
-    currentSortingOrder += maxZSize;
-    isoNode->set_z_index(currentSortingOrder);
-    currentSortingOrder += 1;
-}
-
-Array IsometricMap::getPositionableBehind(IsometricPositionable *isoNode) {
-    Array isoNodes;
-    const Array &children { get_children() };
-    for (int i = 0; i < children.size(); i++) {
-        auto *positionable = cast_to<IsometricPositionable>(children[i]);
-        if (positionable && positionable != isoNode) {
-            if (IsometricServer::getInstance().doHexagoneOverlap(isoNode->getHexagoneCoordinates(), positionable->getHexagoneCoordinates())
-            && IsometricServer::getInstance().isBoxInFront(isoNode->getAABB(), positionable->getAABB())) {
-                isoNodes.append(positionable);
-            }
-        }
-    }
-    return isoNodes;
 }
 
 Array IsometricMap::getFlattenPositionables(const Vector3 &offset) {
